latest

+++ traj_pred/vis_traj_pred.py
@@ -13,16 +13,16 @@ from motion_infiller.data.amass_dataset import AMASSDataset
 from traj_pred.utils.config import Config
 from traj_pred.models import model_dict
 from lib.utils.tools import worker_init_fn, find_last_version, get_checkpoint_path
-from lib.utils.vis import hstack_videos
+from lib.utils.vis import hstack_videos, hstack_3videos
 from lib.utils.torch_utils import tensor_to
 
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--cfg', default='trajv2_s300_ddec_heading_aa')
 parser.add_argument('--split', default='test')
-parser.add_argument('--seq_len', type=int, default=300)
+parser.add_argument('--seq_len', type=int, default=800)
 parser.add_argument('--num_seq', type=int, default=2)
-parser.add_argument('--num_motion_samp', type=int, default=3)
+parser.add_argument('--num_motion_samp', type=int, default=5)
 parser.add_argument('--multi_step', action='store_true', default=False)
 parser.add_argument('--ntime', type=float, default=2e6)
 parser.add_argument('--gpu', type=int, default=0)
@@ -40,7 +40,9 @@ cfg.seq_sampling_method = 'length'
 seq_len = cfg.seq_len if args.seq_len == -1 else args.seq_len
 
 # val datasets
-test_dataset = AMASSDataset(cfg.amass_dir, args.split, cfg, training=False, seq_len=seq_len, ntime_per_epoch=int(args.ntime))
+# J: make it equal
+test_dataset = AMASSDataset(cfg.amass_dir, args.split, None, training=False, seq_len=seq_len, ntime_per_epoch=int(args.ntime))
+#test_dataset = AMASSDataset(cfg.amass_dir, args.split, cfg, training=False, seq_len=seq_len, ntime_per_epoch=int(args.ntime))
 test_dataloader = DataLoader(test_dataset, batch_size=1, num_workers=0, pin_memory=True, worker_init_fn=worker_init_fn)
 
 # checkpoint
@@ -54,7 +56,7 @@ traj_predictor = model_dict[cfg.model_name].load_from_checkpoint(model_cp, cfg=c
 traj_predictor.to(device)
 traj_predictor.eval()
 
-visualizer = SMPLVisualizer(generator_func=None, distance=7, device=device, sample_visible_alltime=True, verbose=False)
+visualizer = SMPLVisualizer(generator_func=None, distance=17, device=device, sample_visible_alltime=True, show_smpl='all', verbose=False)
 
 for i, batch in enumerate(test_dataloader):
     if i >= args.num_seq:
@@ -63,30 +65,61 @@ for i, batch in enumerate(test_dataloader):
     print(f'{i}/{args.num_seq} seq_name: {batch["seq_name"][0]}, fr_start: {batch["fr_start"][0]}')
     
     batch = tensor_to(batch, device)
+    seq_name = batch["seq_name"][0]
+
+    try: #J: make sure we already generated sequences
+        mojoResults = torch.load(f'../MOJO-cap/results/{seq_name}.pt')
+        for key in mojoResults.keys():
+            mojoResults[key] = mojoResults[key].float().cuda()
+
+    except Exception as e:
+        #print(e)
+        continue
+
     output = traj_predictor.inference(batch, sample_num=args.num_motion_samp, recon=True, multi_step=args.multi_step)
     
+    # J: need to save for evaluation
+    torch.save(output, f'out/vis_traj_pred/{seq_name}.pt')
+
+    # betas are always 0 in the visualiser
+    mojoResults['shape'] = output['shape']
+    mojoResults['frame_mask'] = output['frame_mask']
+    mojoResults['seq_name'] = output['seq_name']
+
     prefix = 'multistep_' if args.multi_step else 'singlestep_'
-    vid_name = f'out/vis_traj_pred/{cfg.id}/v{version}_{cp_name}/{args.split}/{prefix}seq_len_{seq_len}/sd{args.seed}_s{i}_%s.mp4'
+    vid_name = f'out/vis_traj_pred/{cfg.id}/v{version}_{cp_name}/{args.split}/{prefix}seq_len_{seq_len}/sd{args.seed}_s{i}__{seq_name}_%s.mp4'
+
+    # save MOJO-cap
+    visualizer.save_animation_as_video(
+        vid_name % 'MOJO-cap', init_args={'smpl_seq': mojoResults, 'mode': 'sample'}, window_size=(1500, 1500), cleanup=True
+    )
+
     # save GT
     visualizer.save_animation_as_video(
         vid_name % 'gt', init_args={'smpl_seq': output, 'mode': 'gt'}, window_size=(1500, 1500), cleanup=True
     )
-    # save recon
-    visualizer.save_animation_as_video(
-        vid_name % 'recon', init_args={'smpl_seq': output, 'mode': 'recon'}, window_size=(1500, 1500), cleanup=True
-    )
-    hstack_videos(vid_name % 'recon', vid_name % 'gt', vid_name % 'sbs', verbose=False, text1='Recon', text2='GT', text_color='black')
 
-    if traj_predictor.stochastic:
-        # save sample
-        visualizer.save_animation_as_video(
-            vid_name % 'sample', init_args={'smpl_seq': output, 'mode': 'sample'}, window_size=(1500, 1500), cleanup=True
-        )
-        hstack_videos(vid_name % 'sample', vid_name % 'gt', vid_name % 'sample_sbs', verbose=False, text1='Sample', text2='GT', text_color='black')
-        os.remove(vid_name % 'sample')
+    # save sample
+    visualizer.save_animation_as_video(
+        vid_name % 'GLAMR-sample', init_args={'smpl_seq': output, 'mode': 'sample'}, window_size=(1500, 1500), cleanup=True
+    )
+
+    #hstack_videos(vid_name % 'recon', vid_name % 'gt', vid_name % 'sbs', verbose=False, text1='Recon', text2='GT', text_color='black')
+    hstack_3videos(vid_name % 'GLAMR-sample', vid_name % 'gt', vid_name % 'MOJO-cap', vid_name % 'combined', verbose=False, text1='GLAMR', text2='GT', text3='MOJO-cap', text_color='black')
+
+    if False: # J: currently not needed
+        if traj_predictor.stochastic:
+            # save sample
+            visualizer.save_animation_as_video(
+                vid_name % 'sample', init_args={'smpl_seq': output, 'mode': 'sample'}, window_size=(1500, 1500), cleanup=True
+            )
+            #hstack_videos(vid_name % 'sample', vid_name % 'gt', vid_name % 'sample_sbs', verbose=False, text1='Sample', text2='GT', text_color='black')
+            hstack_3videos(vid_name % 'GLAMR', vid_name % 'gt', vid_name % 'MOJO-cap', vid_name % 'combined', verbose=False, text1='Sample', text2='GT', text_color='black')
+            os.remove(vid_name % 'sample')
 
     os.remove(vid_name % 'gt')
-    os.remove(vid_name % 'recon')
+    os.remove(vid_name % 'GLAMR-sample')
+    os.remove(vid_name % 'MOJO-cap')
